Unembed Players from Users.
Having Players embedded in a User was proving more trouble than it was
worth. Now with Player as it's own entity the relationships between the
objects are simpler to work with.

Also renamed the Player field game to for_game as the it would get
confusing (and may not work?) to have a singular named attribute and
a plural named attribute of the same name but different things. Will
probably do something similar for Game.game.