Handle logline data via dcss_henzell.
Like so many other things it is hardcoded all over the shop but in
general it works. It makes use of sqllog.pl in dcss_henzell to do the
heavy lifting of turning logfile lines into data. This requires a
minor change to sqllog.pl which I have in a fork.

Importing the data is pretty nippy at about 100 lines a second on an
idle hosted VM with mongod running.

The web frontend also handles the sparse logline data and currently
just links to the full morgue.