Sort resistances as they are in the game.
It's the order people are familiar and it's easy enough to do so why
not. Also hand-wave rAcid as handling its absence is fiddly.

But the real impetus was that Mongoid was silently discarding updates
because the resistances contained dots in them. Things that don't make
sense about that are a) fields with dots can be created in the first
place and b) the update failure was silent.