Make log watcher more flexible.
Now it can ostensibly follow any of the logs and keep its state of where it
was at. This makes having it run across multiple versions and server
considerably simpler. Also store the source of the logfile games so further
ruminating can be done e.g the morgue can be pulled down from the appropriate
server.