Simplify player page related code and include nemeses.
So now the Player owns its own map/reduces and there's less code doing
more. The accummulators were dropped because a) they were fiddly and
b) relied on morgues data being in place which isn't always the case
with logfile based data.

Making use of Hashie in a small way for a start, would like use it in
anger on the morgues for example.

Also fixed a bug where the player's XL was actually being reported as
the last level they were seen on (originally noticed by rebthor but I
only twigged on what the problem was earlier).